refactor(frontend): normalize templates routing to layout + index
- Convert templates.tsx into a pure Outlet layout route
- Move list page logic into templates/index.tsx using the real
  useTemplates/useAwsClient API (was duplicated after rebase)
- Wire template card click to the detail page ()
